refactor: remove `Mixin` from `SqsFifoPartialProcessorAsync`
